Restrict overload stress tests to release
Only run our overload performance stress suites in Release. The suites guard against
overload bugs which previously caused the compiler to virtually hang.

The situation we're guarding against is algorithm flaws in the compiler that would
easily reproduce even in release code. This is evidenced by the fact that customers
found this code on released compilers :)

This stress though causes an unacceptable increase in time to run the suites in
DEBUG mode. The semantic tests jump from 3-4 minutes of execution to 12-13. There is
no extra value gained by running these in DEBUG hence we're disabling them
in that context.

closes #25004