Make AbstractProjectCodeModel.GetCodeModelCache private
This was being used directly in a lot of places, but I think it's a bit
eaiser to follow object relationships if all creation methods go
through AbstractProjectCodeModel, and from there might get forwarded.

A fair number of checks were being made of whether the CodeModelCache
was null; I don't believe these make sense. Once the underlying
_codeModelCache type is made non-null, it is never set to null, and the
only way it could fail to create in the first place is if the project
was never in the ProjectTracker. Given we only produce these once we
already have a live AbstractProject, that's not possible.