Fix netcoreapp2.0 use in TestUtilities
The Roslyn code base defaults to enabling the
`<DisableImplicitFrameworkReferences>` property and instead specifies manual
references. This is mostly an artifact of our code base evolution from
the project.json days where implicit references weren't an option.

This means that even though TestUtilities specified netcoreapp2.0 as a
TF it was still mostly targeting netstandard1.3. This is because
implicit references weren't expanded and we were manually adding only a
subset of the available packages.

Enabled implicit references here, conditioned the explicit ones to
netstandard1.3 and now we can remove a number of `#if` uses.

closes #25228