Account for Mutex ctor throwing
This changes the compiler server to be resilient in the face of the
Mutex constructor throwing an exception. It's documented as being able
to throw various exceptions but we never accounted for it. Had a real
world case pop up when using certain Docker containers on Linux.

In the case the Mutex ctor throws an exception just fall back to normal
csc / vbc execution. The build will complete just fine here, it just
won't have the performance boost of the compiler server.

closes #24124