Delete VisualStudioSetup.Next
This project originally existed to support a Dev14 vs. Dev15 split,
which we no longer version that way. This project had it's contents
moved out a few minor updates of Dev15 ago, and is now empty, only
so that way if you had an older VS Update F5 would still work. Now
we require 15.6 as a base drop, which also has this containing nothing,
so we can just delete this.