Sourcebuild updates
This moves a few of the patches from sourcebuild into the Roslyn source proper.
We now condition on DotNetBuildFromSource in some of our imports, and ensure
that we're not targeting netframework when building for it. We also add a sln
for SourceBuild, SourceBuild.sln, which includes the minimum projects that it
needs.