Add empty Directory.Build.rsp files
When msbuild.exe is invoked it will search the current directory looking
for a Directory.Build.rsp file to provide additional command line
arguments. This search follows the same logic as for
Directory.Build.props files and has the same type of implications
for build reliability.

Adding a dummy empty file for both our standard build and tests.