Enable implicit framework references
Implicit framework references were disabled at the time we moved over to
the new SDK due to some remanants of project.json in our build. Those
have all been cleaned up now and hence moving us over to using implicit
framework references.

Not having this enabled meant the API surface we were targeting in our
projects was not actually what was available in the project target
framework. Instead it was the intersection of the target framework and
the set of NuGet packages we reference. This meant that code which
should work simply wasn't due to the API surface not being correct.