Fix type error in fixed buffer codegen
When generating a pinning temp for fixed statements on
fixed buffers, the type for the pinning temp is ref to
the type of the fixed field. However, in our lowering
the type of a fixed buffer field is a pointer type, while
what we actually want for the pinning temp is a ref to
the underlying type. This change fixes that scenario.

Fixes #26335