Migrate VSTS YAML to dev15.7.x
This copies our VSTS YAML file into dev15.7.x. This changes nothing
about our official build in this branch. It just seeds it with the
correct value and lets the merge bots move it to the rest of the trees.

The official switch over for 15.7 will happen at a later time and will
be very deliberate.