Exclude Microsoft.Composition as something we build a dependency map for
This creates ironically the reverse effect you might imagine:
CompilerExtension contains this package now, which would cause it to be
added to the dependency map; this then causes the package to be
excluded from the insertion package, where it is supposed to be. Since
CompilerExtension and VS are (in some sense) separate targets, presence
in the former shouldn't impact the other.