Add diagnostic options to SyntaxTree (#25964)
This change allows the user of a Compilation to configure warning
reporting on a syntax tree level. There is a new public API, 
DiagnosticOptions, on SyntaxTree which allows a user to retrieve
these options in a map from diagnostic ID to severity.

In general, these options take precedence over the global diagnostic
options, but are overridden by pragmas. In C# the global warning
level can also suppress these options.