Stop calling IVsHerarchy.ParseCanonicalName if we don't need it
We were calling IVsHierarchy.ParseCanonicalName to get the item ID of
a file, to then turn around and use that to get the folder names. In
some cases, we already had the folder names so this was just silly.

Note: I update a compatibility overload for TypeScript here. They had
since moved off of the compatibility overload, so updating it and
repurposing it is safe to do.