Create life-time regions for flow captures in Control Flow Graph. (#28100)
Create Control Flow Graph regions to designate life-time of flow captures in a Control Flow Graph. A region should start as soon as a capture is needed to accurately represent the order of evaluation within operations involving control of flow. However, it should not start too soon. For example, if there is a branch that represents an execution path that bypasses the use of a capture completely, that branch should not be included into the region for the capture. Also, the region should end as soon as the last operation referencing the capture is reached.