No longer generate packages for teams that we now consume proper packages for
The editor ones are now provided to us as proper packages which we need
to consume because we need to be in sync with other teams (like VS for
Mac). Some of these I'm also deleting because we're now using real
packages anyways so this only would have caused extra confusion.