Have DocumentState hold onto a DocumentInfo.DocumentAttributes
DocumentInfo is a type that holds onto some basic information about
a document, including it's TextLoader. The fact it's a TextLoader
is a problem, so 446629fa0fb5e261dc80b89c44edfc2d7f9dec94 introduced
a type that held onto just the info without the TextLoader.
DocumentState was holding onto a DocumentInfo and throwing out the
TextLoader manually to avoid leaks. That code was written prior to
the introduction of DocumentAttributes, but now that we have it we can
just simplify everything.