Fix IntroduceVariable when the textSpan is whitespace
If the textSpan is a span of whitespace that is itself surrounded by
whitespace, trying to recover by getting the next token and creating a
new textSpan will result in a gold bar. This change prevents the gold
bar by checking the proposed newStart position and returning if it is
past the end position.

This fixes #27949 and fixes #28665.